[CP-1070] Move release logic to update module
