+++ packages/app/src/update/constants/ipc-event.enum.ts
@@ -1,0 +1,9 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export enum IpcUpdate {
+  GetDevelopmentRelease = "get-all-development-releases-request",
+  GetProductionRelease = "get-production-release-request",
+}
