+++ packages/app/src/update/helpers/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export * from "./map-to-release.helper"
+export * from "./os-releases-manager.helper"
