+++ packages/app/src/update/types/github-release.type.ts
@@ -1,0 +1,21 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export interface GithubReleaseAsset {
+  content_type: string
+  size: number
+  url: string
+  name: string
+  browser_download_url: string
+}
+
+export interface GithubRelease {
+  tag_name: string
+  created_at: string
+  published_at: string
+  draft: boolean
+  prerelease: boolean
+  assets: GithubReleaseAsset[]
+}
