+++ packages/app/src/update/types/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export * from "./github-release.type"
+export * from "./release-manifest.type"
+export * from "./release.type"
